CC-2522 ConceptTree: if a cycle was broken, don't do resorting
Further to the previous commit, there's more to be done about avoiding
NPEs after breaking a cycle.

It seems it's possible that after breaking a cycle, an NPE can be
thrown during resorting by NotationComparator.compare().

So don't attempt resorting, if we broke a cycle.

Add Joel's test case that exercises both this situation and the one
dealt with in the previous commit.