CC-2522 ConceptTree: omit error messages for non-visited nodes
When depth-first search (DFS) of the resource graph is completed, we
see if there are any non-visited nodes.  If there are any, we have a
cycle. We choose a non-visisted node, make it a root, and continue the
DFS.

At the point we choose a non-visited node, we log an error message for
that.  The fact that we log that error message is a working out of a
principle of recording an error message as soon as we discover each
error in the data.  So we "should" add an error here. But the node we
choose for cycle-breaking may not itself be in any cycle, so the error
message may be "collateral", and not-very-helpful for the user.

But precisely because when we detect the cycle, we continue the
search, we will _eventually_ discover a back edge, and we will log an
error message for that back edge.

So, in this case, we don't _need_ to log the error message for
detecting a non-visited node.

Decision: we trust ourselves that we will indeed subsequently
find the cycle, and add an error message for the back edge.

Comment out the definitions and all of the uses of the two error
messages for the detection of a non-visited node.

RVA-45