CC-1524 Have language tag parsing API methods return canonical form
The language tag parsing API methods should return the canonical form
of the tag(s) being parsed, but they didn't.

Change the return type of parseLanguageTag to LanguageDetails so that
the canonical form can be included in the result.

Have parseLanguageTags return canonical forms, not the forms provided
in the request.

RVA-41