CC-1524 Have parseLanguageTags return a list element for every input
While coding the Portal's pre-filling of languages, it became clear
that the parseLanguageTags method should return a list of responses
that corresponds 1:1 with the list of tags in the input, to facilitate
matching up the response with the request.

So, have parseLanguageTags include a list element for every tag in the
input list, but for invalid tags, don't add attributes.

RVA-41