CC-2774 ConceptTree: iterate only over collections, where possible
The method populateOrderedCollectionMembers() has an iteration over
all master resources, whereas there's a perfectly good Map
collectionMap, containing just the collections, that should be used as
the basis of the iteration instead.

Well, except that collectionMap is incorrectly populated, because of a
missing "break" in a switch statement: all resources marked as
concepts are added as well!

So, the one existing occurrence of an iteration over
collectionMap.values(), in freezeResources(), is currently "broken".
(But this was not a problem in practice, because if a concept ever had
scaffold members, this would indicate a "type error" that would
already have been picked up by the processing done by
handleStatement().))

So, insert the "break" where it's needed in the switch.

Since I had to look at the handling of the skos:member property in
handleStatement(), I noticed the method now-wrongly named
addBroaderMemberForCollection(). Take the opportunity to rename the
method addBroaderMemberForCollection() to
addMemberForUnorderedCollection().

Change the iteration in populateOrderedCollectionMembers() to iterate
over collectionMap.values().

RVA-173