CC-2774 ConceptTree: delete topmostResources; use roots directly
The method populateRoots() very misleadingly populates the Set
topmostResources. But then the elements of that Set are copied one by
one into another Set, roots. This is a remnant of an early stage of
development; there's now no reason to do this.

Remove topmostResources, and insert values into roots in the first
place.

There's a side effect: the iteration over the elements of roots can
now happen in a different sequence, which affects where during the
depth-first search that ResourceRefs are created and inserted into the
forest. So, update the expected results of the affected test cases.

RVA-172