CC-1524 Explain the possibility of duplicate resolved languages
The method that resolves language tags into human-readable form is no
longer a one-to-one function.  For example, we now support both "en"
and "en-CA", which both resolve to "English".

When creating a Solr document, we add a resolved language value for
each of the vocabulary's languages. So the value of the language field
may now contain duplicates.

This doesn't affect Solr search result scores or facet counts.
Nevertheless, document this for future reference.

RVA-41