CC-1524 Remove the possibility of duplicate resolved languages
Change of mind from the previous commit.

The method that resolves language tags into human-readable form is no
longer a one-to-one function.  For example, we now support both "en"
and "en-CA", which both resolve to "English".

When creating a Solr document, we add a resolved language value for
each of the vocabulary's languages. So the value of the language field
could now contain duplicates, if we took no action to remove them.

Duplicate resolved language values in the Solr document don't affect
Solr search result scores or facet counts.

Nevertheless, remove duplicates.

RVA-41