Upgrade Checkstyle to release 9.1
Having upgraded my Eclipse, the Eclipse Checkstyle plugin
now uses a later release of Checkstyle. So might as well
upgrade in the repo.

Upgrade Checkstyle to release 9.1, and patch the configuration and
code to match.

This happened to catch two places where I had used a normal comment
instead of a Javadoc comment. I looked for other such places, and
found one, which I also fixed.