CC-2809 Set timeout for network connections for PoolParty backups
In June, during routine checking of production RVA, found three hung
processes which are instances of our nightly PoolParty backup
program. (The earliest one dated back to March.) Hung here means not
consuming much CPU, but each one was holding on to 1 GiB of swap. Not
good; if this had happened a fourth time, it would have consumed all
remaining swap.

This has happened before a few times, but I just killed the process
and moved on. But that was wrong; I didn't previously notice the
implications for swap. We probably do have to fix this to prevent a
failure.

Preliminary investigation of the cause found
https://stackoverflow.com/questions/32262946/java-periodically-hangs-at-futex-and-very-low-io-output
. The symptoms (i.e., the use of futex_wait) match, but none of the
offered diagnoses seem to match.

The stacktraces show a thread waiting on a socket read.
So try adding timeouts to the Jersey socket read. This may help:
https://stackoverflow.com/questions/19543209/how-to-set-the-connection-and-read-timeout-with-jersey-2-x

Hmm, doc for ClientProperties.READ_TIMEOUT says "Read timeout
interval, in milliseconds. Default value is 0 (infinity)". So
definitely worth a try.

Add a test case using WireMock.

RVA-214