CC-2625: Use PoolParty post-version-7.0 API
PoolParty 7.1 made backwards-incompatible changes to the project
export API method:

* The HTTP GET method is no longer supported; clients must use POST.
* The parameters are no longer specified as query parameters; they
  must be provided in the body of the POST, in JSON format.
* The API documentation states that the prettyPrint parameter is now
  required. (In fact, it seems that's not true. We'll specify a value
  anyway: the default, which was, and is, false.)

In commit defc656506be781591e979b7de6401a2bbe721b5 we made a hotfix
change to specify version 7.0 of the PoolParty API explicitly in the
API path, so that our code continued to work.

Version 9.0 of PoolParty has been announced, which brings closer the
eventual withdrawl of version 7.0 of the PoolParty API.

So we now undo that change, and instead make the changes in our own
code to cope with the change made to the export API method.

RVA-231