CC-2905 ResourceDocs: trim content to fit in a string field
The resources Solr schema stores vocabulary content in fields of type
string. There's a limit of 32766 characters on such fields. We've now
received a report of a user who published content which exceeds that
limit.

Trim content so that it fits.

Possible future work: see if it's possible to do without these string
fields.

Oops: the test data for this test was configured to use content from
the ConceptTree test data. This was very confusing: the input test
data file test-data1.ttl in the testResourceDocsTransformProvider1
directory wasn't being used at all. Rename it as test-data2.ttl and
use it as a test.

Of course, also add a test for this change.

SD-1540906