CC-2905 ResourceDocs: trim content by length in bytes
In the first commit for this change, I wrote: "The resources Solr
schema stores vocabulary content in fields of type string. There's a
limit of 32766 characters on such fields."

But that's wrong. The limit is 32766 bytes of UTF-8, not of
characters. There's a difference!

Redo the implementation of trimming based on bytes of UTF-8 instead of
characters.

SD-1540906