CC-2917 Add helper method to set sequence values
We're going to be adding a test that needs to create versions with
specific version IDs. The way to do that is to set the value of the
sequence underlying the id column of the version_ids table.

So add a helper method that allows us to set the value of any
sequence.

RVA-257