CC-2917 Add tests of task/subtask ordering, disabled for now
Add two tests of the ordering of execution of tasks and subtasks
associated with a single updateVocabulary operation.

The first test creates a vocabulary with one version, and then deletes
and adds the same version (i.e., by specifying a version element
without an id, but with the same title as the existing version) within
the same updateVocabulary operation. The underlying Sesame repository
and SISSVoc spec file should exist after the update, but they don't,
because the tasks are executed such that the delete subtasks are done
after the insert subtasks.

The second test swaps the slugs and titles of two existing
versions. This is even worse than the previous test, because now
there's a cyclic dependency between the subtasks of the two tasks. All
of the deletes across both tasks need to be done before all of the
inserts, and that doesn't happen, and can't happen, because the
TaskRunner operates on one task at a time, and so does all of the
subtasks of one task to completion before starting the subtasks of the
other.

If enabled, the tests would currently fail. So disable them for now.

RVA-257