CC-2917 Add new task runner to resolve inter-task dependencies
Add an alternative way of running tasks. You can now run just the
subtasks of a task that have negative priority, and you can
subsequently run the remaining subtasks.

Use this capability in VersionModel. If there is more than one task to
be run, then first run all of the negative-priority subtasks of all
tasks, and then go back and run all of the remaining subtasks.

This removes all inter-task dependencies, including cycles ... we
hope.

Enable the two tests for this, that were previously disabled.

RVA-257