CC-2917 Minor refactoring of adminApi method for new task runner
In the new adminApi method that allows users to run a set of tasks for
one vocabulary, do a minor refactoring: hoist the assignment of the
vocabulary variable out of a loop.

Add some explanatory comments about the resetting of subtask status
values, including to the constructors of the TaskInfo class.

RVA-257