CC-2940 Don't display "logic error" message, and add some debugging
The altLabel processing would emit "Logic error in addAltLabel" for
the test data. But I'm now fairly happy that the code does not have an
error in this case, so I'm removing that message and adding a comment
explaining why there's no error in the code.

Nevertheless, the code isn't working as expected on testing in devl,
so temporarily add some debugging output.

RVA-270