CC-2940 Copy altLabels to deputy instances
It's been a long time since I worked on this code, and I completely
missed the fact that we "manually" copy attribute values into
deputy instances.

Since I hadn't done that, the altLabels don't appear in concepts that
are below concept schemes or collections.

Copy the altLabels and altLabelLanguage attributes into deputies that
are made for concept schemes and collections.

RVA-270