Checkstyle: fix exception for import statements
The project-specific Checkstyle configuration specifies a maximum line
length for Java files with an exception allowed for import
statements.

It turns out that Checkstyle 9 already allowed long lines for import
statements, so that little bit of configuration was redundant.

In Checkstyle 10, the behaviour changes, and our configuration
override now comes into play. But our configuration override was
always wrong: it didn't allow for long lines beginning "import
static".

Fix the configuration override to allow long lines beginning "import
static". This works with Checkstyle 9 (which we're still using) and
Checkstyle 10.