CC-2992 Use rdf:Property in Elda template
While working on this change, I noticed that I made a mistake in the
Elda template. I referred to a type "rdfs:Property", but it should be
"rdf:Property". It seems the mistake is harmless, but let's fix it.
(It seems in my testing that fixing the mistake is also
harmless. Indeed, the way I've used this type seems to be
"documentation-only", i.e., it has no practical effect.)

Do a search/replace of "rdfs:Property" with "rdf:Property".

RVA-280