CC-2941 Sort superseded versions also by slug
In commit 5fc379ef we implemented sorting of a vocabulary's superseded
versions (as returned by the Registry API methods getVocabularyById
and getVocabularyBySlug, and as used by the Portal view page) in
reverse order of release date.

This works correctly.

However, it turns out that at least one of our users can publish
multiple versions on the same day, and in this case, sorting only by
release date is not good enough. See
https://demo.vocabs.ardc.edu.au/viewById/970 for an example.

We need to add a secondary (and maybe also a tertiary) sort key. This
could be the version title and/or the version slug; choose what seems
to work best.

For now, let's try using the version slug as a secondary key.

RVA-274