HELP-12 Revert https to http in DTD for TestNG
It seems we currently can't access testng.org, and this causes the
testng and swagger-json build targets to fail.

Oh dear, I changed the URL of the DTD for TestNG from http to https in
June (commit 4f46605f) to keep Eclipse's own XML validation happy. It
turns out that there's some "magic" in the TestNG source so that if
you specify the http-based version, TestNG uses its own internal
copy of the DTD, rather than going out to the Internet to fetch it.

So best is probably to revert the references back to the http-based
version, even though that URL is no longer correct.

That way, validation against the DTD will (once again) not require
testng.org being reachable.

RVA-290