CC-2882 Overwrite subject labels and notations from database
The CheckVocabularyImpl class calls out to a helper class
SubjectSources, which was committed in toto in April 2017 and has
not been touched since then. The tantalizing aspect of this is
that this class contains a method resolveSubject with method
documentation as follows:

* Resolve a Subject, based on its source and iri values. The
  subject's label and notation are set to the values obtained
  from the subject resolver. If the Subject does not
  resolve (i.e., it can not be found in the subject database), it
  is left untouched.

But this method is never invoked, anywhere in the Registry code!
I must have meant to use it somewhere, at some time ... but it
seems I never did.

If users only ever use the Portal to create and edit
vocabularies, there is no problem, because the Portal fills in
the label and notation attributes with the correct values before
passing them to the Registry.

Now to the problem:

What happens if you create or edit a vocabulary using the
Registry API, bypassing the Portal? Well, you can pass in
incorrect (or even completely bogus) label and notation values,
and the Registry will accept them, and put them into the (Solr)
index.

It seems that this has happened. To be specific, the Registry API
has been used to create over 100 vocabularies on the production
server that have an incorrect notation value. These values have
this in common: they should have a leading zero, but this leading
zero was omitted when sending the metadata to the Registry.

The notations that are wrong have now been fixed in the
production server.

But we now do something with the validation process; we could
either reject incorrect notations, or fix them. I suspect, given
the existing-but-dead code, I intended to silently fix them.

So now, I choose to overwrite the label and notation values from
the database, where they are available.

We take the opportunity to fix a separate defect:

SubjectSources has a flag resolvingSubjectSourcesPopulated which
should be set to true after the RESOLVING_SUBJECT_SOURCES set is
populated. But it was never set to true, so the set was
re-populated with every invocation of subjectRequiresIRI().

Now, we set the flag to true after populating
RESOLVING_SUBJECT_SOURCES.

RVA-295