CC-2531 Use "most suitable" version for indexing
The logic used to determine if a vocabulary is "widgetable" was not
consistent between the Portal search results and view pages.

Old logic:
* Search results: a current version with a SISSVoc endpoint.
* View page: either a current version, or, if there is no current
  version,  a superseded version with a SISSVoc endpoint.

For CC-2941 we sorted superseded versions also by slug, and in the
Portal we now base the presence of the widgetable icon on the view
page on the "most suitable" superseded version if there is no current
version.

We now adopt this logic also when creating the Solr document for
indexing: we choose the "most suitable" version as:
* the current version, if there is one;
* the most recent superseded version (disambiguated by slug), if
  there's no current version.

RVA-297