CC-2466 Add DISTINCT to /concept?labelcontains's SPARQL query
This is the first in a series of a commits to include upstream changes
to SISSVoc. This one comes from SD-10382:

The /concept?labelcontains endpoint can return duplicates, because, by
default, the underlying SPARQL query says "SELECT", not "SELECT
DISTINCT". The "DISTINCT" keyword _is_ added if you specfy a _sort
parameter.

Change the definition of the endpoint so that "DISTINCT" is always
applied.

See also the upstream reports:

* https://github.com/SISS/SISSVoc/issues/5
* https://github.com/epimorphics/elda/issues/173

RVA-298