CC-2466 SISSVoc: add api:base setting
The third in a series of a commits to include upstream changes to
SISSVoc.

Set the value of api:base.

In my own testing, the value must end with a slash, so add one.

Without a trailing slash, generated URLs for example request paths
shown at the /api-config page are wrong: they omit the last path
component of the api:base setting. (I.e., in our case,
/repository/api/lda gets stripped to /repository/api/.  But when
api:base is set to /repository/api/lda/, the generated URLs are
correct.)

See upstream commit 5ea5033, and upstream issue
https://github.com/SISS/SISSVoc/issues/18 .

RVA-298