CC-2704 Re-run Sesame import on PoolParty unharvest
Consider a vocabulary that is based on a PoolParty project, and that
has an existing version such that:

* do-poolparty-harvest="true";
* do-import="true";
* the version has a file access point.

Now update the version, but set do-poolparty-harvest="false".

In this case, a Sesame import subtask must be run, to clear out the
concept data that was previously harvested, and to re-import just the
concept data from the file access point.

But that is not happening; no Sesame import subtask is run.

We already had a test for this, which was disabled:
RegistryModelWorkflowTests.testApplyChangesCurrentVoVREVeAPVAW12.

Some of the logic has now been moved around, so that an import task is
scheduled in this case. The test is now enabled, and passes.

RVADEV-13