CC-2581 Disallow vocabulary and version titles that give empty slugs
It's important to check that if a slug is being generated from a
title, that the resulting slug is not an empty string.

For example: you can create a vocabulary by specifying a vocabulary
title "?" and not specifying a vocabulary slug. Until now, this was
allowed, and the resulting vocabulary had a vocabulary slug which was
an empty string.

It turns out that you can't then update that vocabulary using the
Portal: when you try to update it, because the slug is an empty
string, the portal omits it in the update request to the registry, and
validation fails.

In a similar way, it was also possible to create a version by
specifying a version title "?", and this also gave a version slug
which was an empty string.

(You could then update such a version, because the version validation
doesn't require that you specify the slug on update.)

But now, we add validation in the places where we generate slugs from
vocabulary and version titles to make sure that we never accept an
empty slug.

RVADEV-17